Change Reporter specs to conform to the data that the Differ will eventually output.
1. For a single change, instead of :old_element, :new_element and :common_type,
we now have an :elements hash that has :old, :new, and :common keys. For an
:equal change, each of these hash will be exactly the same. For an :inequal
change, each of these hashes will have the same keys (at least :value, :type,
and for subchanges, :key), but not the same values: values in :common will be
nil if corresponding values in :old and :new are inequal. For a :surplus change,
:old and :common will be nil; for a :missing change, :new and :common will be
nil.

2. For complex changes, the :details collection has changed. Instead of being an
array of hashes, which represent subchanges, it's now an array of arrays, which
represent *hunks* of a certain type, where a hunk contains multiple changes
grouped together as one operation. We do this to support storing the data that
Diff::LCS will produce. So, we might have a :missing hunk that indicates that
multiple values adjacent to each other were removed, and a separate :surplus
hunk that indicates that multiple values adjacent to each other were added.