Add integration with RSpec
Right now, RSpec doesn't support replacing its differ with a custom
implementation. Therefore, we have to resort to monkey-patching some
internal RSpec classes to use SuperDiff instead. While we're at it,
disable some of the coloring that RSpec does, particularly its habit of
styling failures in blood red. We want to be able to show off the diff
and all of its glory and this just gets in the way.