Run tests on Travis under JRuby
This required fixing the tests so that they could run on JRuby. JRuby
seems to not fully support `spawn`, in the sense that even if you
explicitly tell it to redirect stdout and stderr in the child process
using `out: writer, err: [:child, :out]`, it doesn't work: stdout from
the process leaks through. The `childprocess` gem, however, does work,
so we use that instead.