Add ability to disable color
When using `super_diff/rspec` in a project that is tested on a service
like CircleCI, if a test fails, then the test output will contain color
sequences (because the gem produces them via the Csi module). This
commit provides a way to disable color by configuring the gem like so:

    SuperDiff::RSpec.configure do |config|
      config.color_enabled = false
    end

That said, this setting should take effect automatically if the output
stream is not a TTY, just like RSpec does. So everything should "just
work" on CI.