Fix tests so they run even with a global --color setting
If you have global RSpec settings located at `~/.rspec`, and you have
`--color` as one of your settings, and you try to run the `super_diff`
tests locally, the integration tests, which ensure that output is
printed in the correct colors and which work by spawning a new `rspec`
child process for each test, will fail. I'm not exactly sure why this is
happening, but we can get around this by instructing the child `rspec`
process not to use the global configuration settings.