Fix diff from using have_attributes with a hash
The `have_attributes` matcher is not "supposed" to be used against a
hash, but considering Hash is an object, it is still a perfectly valid
use case. Still, typically, such a test will probably fail, and so the
failure should show an appropriate diff. Currently, the representation
of the subject in said diff is very low-level, as the DefaultObject diff
formatter is being utilized (as Differs::ObjectHavingAttributes inherits
from Differs::DefaultObject, and Differs::DefaultObject uses
DiffFormatters::DefaultObject). Instead we should be using whatever
DiffFormatter is appropriate for the type of subject: hash, array, etc.
Changing this is non-trivial, but it is this with which the majority of
this commit is concerned.