Globalize all configuration
Prior to this commit, configuration was overridable at runtime. This is
mainly a holdover from the initial design of this gem, which provided a
EqualityMatcher class that allowed you to produce a diff of two values
if you weren't using RSpec. The EqualityMatcher.call method provides
options that let you supply custom differs, operational sequencers, etc.

There are two issues with this. One is that the RSpec stuff in this gem
has mainly taken over and EqualityMatcher is not as valuable. The other
is that we end up having to pass down custom classes all the way to
places that need to make use of them. This is proving to be irritating
as it's difficult to remember to always do this. This could therefore
lead to bugs.

Because of this, it's a better design to move all configuration to a
global configuration block, `SuperDiff.configure`. This is already the
preferred method when using "super_diff/rspec" anyway, so we're really
just promoting it. (SuperDiff::RSpec.configure still works but will be
deprecated in a future version.)