Fix compatibility issues with rspec-rails
With 0.6.0, requiring `super_diff/rspec-rails` in the context of
`rspec-rails` suddenly started failing. We try to require
`rspec/version` so that we can have access to RSpec::Version. However,
`rspec-rails` does not depend on the `rspec` metagem, so RSpec::Version
is not a constant when using `rspec-rails`. In addition, this was never
caught because it turns out that integration with `rspec-rails` was not
being properly tested.

This commit fixes `super_diff` to reference RSpec::Core::Version (since
we can always assume `rspec-core` has been loaded) and makes sure that
we are testing it so the aforementioned problem does not occur again.