Remove warnings
When running tests locally, multiple warnings were being emitted. A lot
of these warnings have to do with the fact that we are overriding core
methods in RSpec, and whenever you override a method, Ruby feels the
need to inform you. To get around this, we need to use the same trick
that we already use for AugmentedMatcher: we create a module on the fly
and `prepend` it to the class or module in question. Sometimes we might
even need to `prepend` to the singleton class. Also, this commit adds
WarningsLogger so that we can track when this happens again.