Update gemfiles; ignore gemfile.lock files
It seems that following a previous merge, the generated appraisal
gemfiles are now out of sync with the `Appraisals` file. Namely, the
version of RSpec in the "< 3.10" gemfiles are not set to `~> 3.9` and
`pry-byebug` and `pry-nav` are not located in the correct groups. Also,
the lockfiles aren't being properly ignored. Finally, unrelated to the
appraisals changes, for some reason a Bundler config file exists,
presumably added in a previous commit. This has all been addressed in
this commit.

Note that due to updating the development Ruby version to 3.2.0 in a
recent commit, it is only possible to regenerate all of the gemfiles,
and thus reveal any issues with appraisals, by switching to Ruby 2.7
locally before running `bundle exec appraisal generate` and/or `bundle
exec appraisal install`. This is an unfortunate problem that will be
addressed in a future commit.