Use Prettier to lint/format instead of Rubocop (#178)
I use CoC in Vim to integrate with language servers similar to VSCode,
but I've never gotten Solargraph, which incorporates Rubocop, to work.
On the other hand, Prettier has been a pleasant experience to use, and I
love how I've never had to write a single configuration file.

To fix this, this commit:

* replaces Rubocop with Prettier;
* reformats all files to match Prettier's format;
* adds a new step to the GitHub workflow to ensure that all files pass
  the lint step; and
* adds an optional Git hook so that before you push, all files will
  linted

Note that in order to get started with this project, you'll need to have
Node installed. I recommend using `asdf` which can manage not only Node
but also Ruby as well. Then you'll want to run `yarn install`. I've
included this in the Contributing doc as well as in the setup script.