Don't load AR integration if it is unavailable
When writing tests for a project that uses certain parts of Rails, it
should still be possible to say `require "super_diff/rails"`, and
SuperDiff should only load the integrations that correspond to the
available parts. Right now the only two integrations are ActiveSupport
and ActiveRecord. ActiveSupport should always be available, but
ActiveRecord may not be available, so this commit adds a guard around
that integration in particular. It also adds tests for Combustion so we
can test that this guard specifically works for engines where
ActiveRecord is not being loaded.

Co-authored-by: Harry Lascelles <harry@harryl.com>
Co-authored-by: sshaw <skye.shaw@gmail.com>