Enforce use of Bundler >= 2.4
We make use of `install_if` in our appraisals. However, if you aren't
using Bundler 2.4 then you can't generate any appraisal gemfiles.

This commit enforces the correct version of Bundler by adding a check to
the Gemfile itself. This isn't ideal, but we can't simply add `gem
"bundler", "~> 2.4"` to the Gemfile because of the new-ish logic in
Bundler which [automatically switches to the version of Bundler
specified by the `BUNDLED WITH` section in the lockfile][1].
Unfortunately, we can't commit the lockfiles to the repo, so neither
adding the `gem` line nor updating the `BUNDLED WITH` would work.

[1]: https://bundler.io/blog/2022/01/23/bundler-v2-3.html