Update dev Ruby; drop old Rubies and Rails
We support versions of Ruby and Rails which are or are about to be EOL:

- Ruby 2.5 and 2.6 are already EOL, and Ruby 2.7 will be EOL starting
  April 1st, 2023 ([1]).
- Rails 5.0, 5.1, and 5.2 are already EOL ([2]).

To address this, I've removed support for these from the Appraisals file
as well as the CI matrix. I've also simplified the Appraisals file and
CI matrix, as we were excluding combinations of Rails and RSpec versions
that didn't need to be excluded.

I've also updated Appraisal to the latest commit within the repo.
For some reason, if you remove `gemfiles` entirely and then re-run
`appraisal install`, Appraisal will complain that a lockfile doesn't
exist. This seems to be fixed in more recent commits. (I tried to fix
this in another commit, but then you need to be able to update _all_ of
the appraisal gemfiles. But to do this, you need to be able to switch to
old versions of Rubies, but I can't do this because those won't install
on new versions of macOS. This commit removes the old Rubies, so that's
unnecessary.)

Finally, I've updated the GitHub Actions workflow to run the lint and
audit jobs first, then run the test job if those succeed.

[1]: https://endoflife.date/ruby
[2]: https://endoflife.date/rails