Only lint changed files when pushing branch
The Git hook that's currently configured will lint all files in the
project when a branch is pushed. Linting takes a long time, so to speed
this up, only lint the files that have changed in the branch. Also, to
make the feedback loop even faster, fix those files instead of just
linting.

Also rename the package script to create the Git hooks from
`simple-git-hooks` to `setup-git-hooks` (so if we decide not to use
`simple-git-hooks` in the future we don't have to rename it again) and
update `bin/setup` script to run this script.

Finally, replace `simple-git-hooks` with `husky`. Our hook is now a
little bit more complicated as we need to figure out which branch the
current branch is based off of, find the files changed in the branch,
and run Prettier accordingly. Being able to just commit the hooks to the
repo as Husky allows us to do makes it possible to guarantee that we are
doing this.