Respect extra_failure_lines in RSpec metadata (#208)
RSpec tests can have associated metadata, which is just a hash of
content. One magic metadata property is `extra_failure_lines`. If this
property is set, then RSpec will read its content and include it in the
output it prints if the test fails.

This property particularly comes into play in RSpec system specs, a
wrapper around Rails system tests. If a system test fails, a screenshot
is automatically taken of the page (provided the driver supports it) and
a message is printed which details the path to the screenshot. RSpec's
wrapper will capture output generated while the test is running and then
store it in `extra_failure_lines`. By doing this, output (including the
screenshot path, but potentially other information as well) is
effectively moved to the spec failure message.

This gem, as it patches key parts of `rspec-rails`, doesn't support
`extra_failure_lines`, so anything in this metadata property is
effectively discarded. This commit restores support.

---

Co-authored-by: willnet <netwillnet@gmail.com>