Remove instance_eval from InspectionTree (#210)
The InspectionTree class defines methods that create nodes in the tree.
Some of these methods take a block which is `instance_eval`'ed inside of
a new subtree. This use of `instance_eval` may seem to create a very
clean and easy to use API on the surface, but it quickly breaks down if,
as you are defining your InspectionTree, you need to extract methods
inside of your InspectionTreeBuilder and then reference them. _Not_
`instance_eval`'ing creates footguns, but it also removes the magic and
mystery around how InspectionTree works.