Fix Zeus
The Zeus config has been broken since integration tests for Rails
engines were added. It appears that the Zeus boot step responsible for
loading Combustion was passing an empty array for the modules to
initialize. Outside of Zeus, the integration test customizes this array,
but inside Zeus, that never happens, and there isn't a way to do that.
So, considering that we only have one engine-specific integration test,
I removed the customization option and hardcoded the list of modules,
renaming the boot step and corresponding command to match. I then
corrected the test accordingly.