Fix match_array to truly accept non-array (#213)
Previously, it was noticed that `match_array` raised an error when given
a single, non-array argument as opposed to an array (which is the
documented usage). An attempt was made to fix this; however, the
conditional added to `match_array` only checked that the single argument
was a string. This commit matches the original implementation of
`match_array` by accepting any type of non-array, not just a string.

---

Co-authored-by: wata_mac <watassbass@gmail.com>