Improve raise_error messages
This commit makes some minor changes to the `raise_error` matcher to
make the failure messages slightly more friendly and consistent with
the overall philosophy of matchers in this repo.

Here are the current messages in comparison with the new ones:

- In cases where the block raises no error, the wording changes
  from "Expected block" to "Expected exception-free block", and
  "exception-free block" is highlighted in the actual color. This is to
  emphasize the actual behavior vs. the expected behavior and matches
  other matchers.
- In cases where only an expected error class is provided, the wording
  changes so that instead of inspecting the expected error, the failure
  message says "a kind of" followed by the class name.
- In cases where a message is provided, the expected class and message
  are separated in the failure message.
- In cases where a regexp message is provided, the wording changes from
  "with message #{message}" to "with message matching #{message}".

Additionally, the integration tests for `raise_error` have been reviewed
and missing tests have been filled in.