Add ability to inspect all describable matchers (#219)
In a previous commit, a new inspection tree builder was added which was
designed to handle built-in RSpec matchers as well as matchers created
via the matcher DSL, by reusing the `description` of the matcher as the
inspection text. However, this inspector did not cover custom matchers
defined from scratch rather than via the DSL. This commit corrects this
by making use of the `is_a_describable_matcher?` method from
RSpec::Matchers.