Add docsite
`docs/` now holds Markdown files that will be published to a docsite
generated by `mkdocs` and hosted on GitHub Pages. Using `mkdocs` allows
us to try out literate programming. This requires Python, so that's been
added to the `setup` script.

Thanks to some new GitHub actions, provided there have been any changes
to the docs, a preview version of the docsite will be automatically
published when a pull request is updated (in which case the URL will be
`https://mcmire.github.io/super_diff/branches/<branch-name>/<commit.id>`),
and a release version will be published when a release PR is merged (in
which case it will be
`https://mcmire.github.io/super_diff/releases/<version>`). Any preview
version that have been previously been created for a pull request will
be deleted when that pull request is closed.

The CONTRIBUTING and ARCHITECTURE docs have been moved into `docs/` and
expanded to not only explain how SuperDiff works but also the RSpec
integration.