Add lint:changed and lint:changed:fix
The Git hook that runs on push will only lint files changed within the
current branch. However, it won't fix anything; to do that, you'll have
to run `yarn lint:fix`. But actually, this runs every file through the
linter, so it's fairly inefficient. This commit adds `lint:changed` and
`lint:changed:fix` so that you can do the same thing that the Git hook
does.