Support the use of primary keys other than "id" in ActiveRecord.
Previously the diffing code made an assumption that "id" would always be
the primary field. This is not always the case. Additionally, the use of
read_attribute(:id) is deprecated in rails/rails@39997a0.

This changes adds support for use of custom primary key fields, and
updates the Person test model to use the "person_id" primary key.

Note that this will almost certainly still run into issues if the model
uses the newly introduced composite primary key type, so this would
require additional changes outside the scope of this fix.