Fix GH workflow so docs deploy on new releases (#245)
Previously, when merging a new release, a new version of the docsite
would not get deployed. Apparently I misunderstood the purpose of the
`github.event.merged` property. I've changed this so that the step to
determine the destination for the docsite (and hence whether the docsite
should be deployed) looks for the `main` branch to be pushed and for
`version.rb` to be updated.

I've also fixed the index page for the docsite so it redirects to the
correct release page.