Remove unused Csi methods
It's hard to keep Csi.color_enabled and
SuperDiff::Configuration.color_enabled in sync. But Csi.color_enabled is
only serving Csi.colorize, which in turn only exists for
Csi.inspect_colors_in, which is apparently unused elsewhere in this
project.

Since Csi is a library private to this project, we can remove them.