Fix hash diffing algorithm (#293)
Fixes #99.

This rewrites the hash diffing algorithm to leverage LCS on the hash
keys, maintaining the goal of preserving the order of the `actual` keys
where possible.

This causes a couple notable changes. Previously, changes to a shared
key's value were sometimes rendered as separate deletes and inserts.
These changes are now always visually colocated as a delete of the old
value followed immediately by an insert of the new value.

Additionally, keys at the beginning of `expected` that are not in
`actual` were previously rendered as deletes at the end of the diff.
These are now rendered at the beginning, reflecting the key ordering
relative to `actual` more accurately.